src/ServerBioStream/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
pillow==11.1.0
setuptools==75.8.0
sqlparse==0.5.3
wheel==0.45.1